Fix issues found by FindBugs: GC_UNRELATED_TYPES http://findbugs.sourceforge.net/bugDescriptions.html#GC_UNRELATED_TYPES

+++ src/main/java/net/onrc/onos/core/devicemanager/OnosDeviceManager.java
@@ -319,7 +319,7 @@ public class OnosDeviceManager implements IFloodlightModule,
 
     @Override
     public void deleteOnosDeviceByMac(MACAddress mac) {
-        OnosDevice deleteDevice = mapDevice.get(mac);
+        OnosDevice deleteDevice = mapDevice.get(mac.toLong());
         deleteOnosDevice(deleteDevice);
     }
